refactor: extract `overridesMap` into a function
